moved formatting functions to interface
